backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.8.2
httpx==0.27.2
aiohttp==3.10.5
python-multipart==0.0.12
python-dotenv==1.0.1
tenacity==8.5.0
azure-storage-blob==12.19.0
azure-identity==1.15.0
